Разбил скрипт на модули. Заменил объявление функций на стрелочное
